backend/requirements.txt
fastapi
uvicorn
nltk
rank-bm25
torch
transformers
python-dotenv
google-generativeai